[020] Address Glenns Comments

+++ src/main/java/org/kainos/ea/api/JobRoleService.java
@@ -4,8 +4,6 @@ import org.kainos.ea.cli.JobRole;
 import org.kainos.ea.cli.JobRoles;
 import org.kainos.ea.client.ActionFailedException;
 import org.kainos.ea.client.JobRoleDoesNotExistException;
-import org.kainos.ea.client.ValidationException;
-import org.kainos.ea.core.JobRequestValidator;
 import org.kainos.ea.db.DatabaseConnector;
 import org.kainos.ea.db.JobRoleDao;
 
@@ -14,12 +12,10 @@ import java.util.List;
 public class JobRoleService {
     private final DatabaseConnector databaseConnector;
     private final JobRoleDao jobRoleDao;
-    private final JobRequestValidator jobRequestValidator;
 
-    public JobRoleService(DatabaseConnector databaseConnector, JobRoleDao jobRoleDao, JobRequestValidator jobRequestValidator) {
+    public JobRoleService(DatabaseConnector databaseConnector, JobRoleDao jobRoleDao) {
         this.databaseConnector = databaseConnector;
         this.jobRoleDao = jobRoleDao;
-        this.jobRequestValidator = jobRequestValidator;
     }
 
     public List<JobRoles> getJobRoles() throws ActionFailedException {
@@ -35,8 +31,11 @@ public class JobRoleService {
         return jobRole;
     }
 
-    public String deleteJobRoleById(int id) throws ActionFailedException, ValidationException {
-        jobRequestValidator.checkJobRoleIdExists(databaseConnector, jobRoleDao, id);
+    public String deleteJobRoleById(int id) throws ActionFailedException,
+            JobRoleDoesNotExistException {
+        if (jobRoleDao.getJobRoleById(databaseConnector.getConnection(), id).isEmpty()) {
+            throw new JobRoleDoesNotExistException("Job role with ID " + id + " does not exist");
+        }
         return jobRoleDao.deleteJobRoleById(databaseConnector.getConnection(), id);
     }
 }
